Moves api into lib
