+++ config/lib/services/text-it.js
