+++ lib/batches.js
@@ -2,7 +2,7 @@
 
 const logger = require('heroku-logger');
 
-const textIt = require('./text-it');
+const textIt = require('./services/text-it');
 const { allSubscribersGroupId } = require('../config/lib/batches');
 
 /**
