+++ lib/middleware/authenticate.js
@@ -3,7 +3,7 @@
 const auth = require('basic-auth');
 const logger = require('heroku-logger');
 
-const config = require('../../config/middleware/authenticate');
+const config = require('../../config/lib/middleware/authenticate');
 
 /**
  * @param {Object} user
