+++ lib/middleware/parseFlowEvent.js
@@ -2,7 +2,7 @@
 
 const logger = require('heroku-logger');
 
-const textIt = require('../../lib/text-it');
+const textIt = require('../services/text-it');
 
 module.exports = function parseFlowEvent() {
   return async (req, res, next) => {
