+++ lib/middleware/sendResponse.js
