+++ lib/middleware/zapier/postZapierWebhook.js
@@ -3,7 +3,7 @@
 const logger = require('heroku-logger');
 const superagent = require('superagent');
 
-const zapier = require('../../../lib/zapier');
+const zapier = require('../../services/zapier');
 
 module.exports = function postZapierWebhook() {
   return async (req, res, next) => {
