+++ lib/services/text-it.js
@@ -3,7 +3,7 @@
 const client = require('superagent');
 const logger = require('heroku-logger');
 
-const { apiToken, testModeEnabled } = require('../config/lib/text-it');
+const { apiToken, testModeEnabled } = require('../../config/lib/services/text-it');
 
 const baseUri = 'https://api.textit.in/api/v2/';
