+++ server.js
@@ -11,7 +11,7 @@ const app = express();
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
 
-require('./api/routes')(app);
+require('./routes')(app);
 
 app.get('/', (req, res) => {
   res.send('hi');
